+++ src/measure.h
@@ -1,6 +1,7 @@
 #include "ifm.h"
 #include "state.h"
 #include "wolff.h"
+#include <stdint.h>
 
 #define NUM_MOMENTS 2
 
@@ -18,5 +19,5 @@ typedef struct { double v[NUM_MOMENTS][NUM_AVERAGES]; } meas;
 
 void meas_reset(meas *m);
 void meas_accum(meas *sum, wolff const *w, state const *s);
-void meas_average(meas *p, unsigned int t_meas);
+void meas_average(meas *p, uint64_t t_meas);
